Wait for VM when retrieving IP address (#1)
When this function is initially called it’s highly unlikely that the target VM has had the chance to start yet. As such, it’s unlikely to have an IP assigned. By using Tart’s built-in `--wait` flag we can give the VM some time which makes IP resolution more reliable.